Add <pid:...> converter for URL routes
This one raises a ValidationError (which will causes routing to proceed
to the next registered route), rather than a kind of HTTPException.
Maybe I’ll change that for the other converters as well at some point.

The name pid: is chosen instead of property_id: to prevent the routes
from getting too long.