Support file names as input for MediaInfo entities
Finding the MediaInfo entity ID is annoying – let’s let users enter
File:… titles instead and look up the page ID for them. We hint at this
option in the placeholder, too (though so far only if the wiki was
selected in the URL).

The placeholder file for Commons is one of the Wiki Loves Pride 2019
runner-up photos. It looks nice and has a perfectly innocuous title.