Show error if entity is not found
Instead of crashing because the 'lastrevid' member doesn’t exist.

The wbgetentities response for missing entities has `"missing": ""`, not
`"missing": true`, even with formatversion=2; this was discussed in
T212069 but there seems to be little interest in fixing it. (Otherwise,
I’d prefer `entity.get('missing', False)` over `'missing' in entity`.)