Make entity_statements() return all statements
Remove the property_id parameter, and instead make it return statements
for all properties – after all, the part to get statements for one
property is regular. We’ll need to be able to get all statements for the
upcoming batch mode, since there we’ll just have statement IDs, without
the property they belong to.