Add batch mode to set rank of statements
This takes a list of statement IDs in plain text (one per line), and
offers three buttons to set the rank of each statement to preferred,
normal, or deprecated. A fourth button, to increment the rank, is
anticipated, but commented out, and the corresponding endpoint is not
added yet. After the edits are done, a results page is shown that lists
all the edited entities and links to the diffs; the name of that
template does not include the qualifiers “list” (as opposed to “sparql”)
or “collective” (as opposed to “individual”), since I expect to reuse it
for all other kinds of batch mode as well.