Simplify “not logged in” check
Instead of testing if the oauth_access_token exists in the session, then
getting an authenticated_session() and asserting it’s not None, just get
the authenticated_session() and test if it’s None to see if the user is
logged in or not. This is shorter and also consolidates knowledge of the
oauth_access_token value in less places.