Fix empty custom edit summaries
If the edit summary field is left empty, it’s still submitted as part of
the form payload, so comparing custom_summary against None in the edit
summary functions is not enough. Also, we should trim it, so that
nonempty but blank (whitespace-only) custom summaries don’t add a colon
at the end of the final edit summary either.