Widen base_revision_id parameter type
When we get the base_revision_id from the form data, it’s a str and not
an int, and since Werkzeug 2.0.1 mypy knows this too and yells at us if
we pass it into an int param. We can actually pass either int or str
into mwapi, so widen the parameter type accordingly.