Work around layout bug in Chromium
Chromium seems to assign most of the table width to the first table
column with a .col-* class. Previously, this meant that the middle
column (list statements) would take up most of the table width, which
looked awful. With this, the wiki + collective/individual column gets
most of the wdith instead, which is still bad, but good enough that I’m
willing to deploy it and not investigate further. In Firefox, this
(correctly, I think) makes no difference, since Bootstrap’s grid system
is arranged around 12 columns and there were 6 columns left.

Also, move those widths to <col>s in a <colgroup>. This makes no
difference in either browser as far as I can tell, but declaring a
.col-* class on a regular <td> (the wiki) feels strange to me.