Add rainbow background to navbar
Surely no user of the tool would object to some more color in their
lives 😊☺️🥰

The colors are defined using CSS variables for ease of tweaking;
eventually this could be “compiled” into direct constants, but for now
let’s leave it like this.

The items in the navbar (other than the tool name) get a dark shadow, to
be more readable against the busier background; additionally, we
brighten them up a bit too.