Only send edited statements to Wikibase API
It looks like Wikibase validates even statements that didn’t change, so
if an item had any statements that are no longer valid (e.g. removed
language codes), it wasn’t possible to edit any statements on that item
with Ranker. Avoid this error by only sending statements we actually
edit to the API. The property_id optimization is removed – in the only
places where we were passing the property ID, we were already only
passing a statements dict with a single property ID anyways.