Format values as HTML instead of plain text
This makes it much easier to work with monolingual text, since the
language name is included now (and the text itself is language-tagged in
the HTML too). The previous reason to use plain text, according to the
commit message of e2e9f6e34a, was that we don’t want to follow links
when clicking (and also, the links are broken anyways); solve that
instead by turning any links into plain spans.

This includes a rebuild of requirements.txt; you can see by the changed
path that this time I ran pip-compile in the source directory.
Previously, I’d adjusted the “to update, run” command (it used to
include the requirements.in path), but didn’t realize all the “via”
lines would also be different. Meh.

In the tests, the escaped HTML looks a bit different, and also we need
to escape a BeautifulSoup warning, because there are no tags in the
markup and it thinks the / makes it look like a URL. (It doesn’t really
look anything like a URL, and I find the warning quite irritating
actually, but meh.)