Clarify documentation of edit functions
I found this quite confusing yesterday; hopefully this helps a bit. (The
assignment `statement_ids = flask.request.form` in the callers is still
confusing, but I can’t think of a way to clarify it concise enough for a
comment there. The idea is that we use the statement ID for the checkbox
input name, so the form data as a whole will contain a statement ID iff
it was checked.)