Add input for reason for preferred / deprecated rank
Part of #5; this implements the feature for the regular “edit” mode, as
well as for the “collective” batch modes. The “individual” batch modes
(list and query) will be done separately, since they’ll require
different ways of specifying the reason.

The <datalist> of suggested reasons was selected by hand, based on the
constraints of both properties as well as queries for the most common
values of each. I’m open to suggestions for changing it.