Add optional reason in query individual batch mode
Concludes #5. In addition to a ?reason variable, we also support
?reasonForPreferredRank and ?reasonForDeprecatedRank, which are only
used if that rank is selected, and take precedence over ?reason; this
could be useful, for example, by binding those two variables to constant
values, then letting the rest of the query determine the rank, and
having Ranker select the appropriate of the two reasons.

I updated the query in the test to cover several cases, using UNION; the
test_query_results are the actual results of that query when I ran it
earlier tonight.

While we’re at it, fix a typo in statement_id_from_uri() (unmarked
f-string) that I noticed while working on item_id_from_uri().