Don’t send edits for unchanged statements
If none of the statements needed to be changed, there’s no point in
sending an edit: it’ll just return a 'nochange' response (but possibly
still bump the rate limit? not sure). Detect these cases before we send
the request instead.

In batch mode, we can include these cases in a separate list in the
output. In non-batch mode, we don’t have the luxury of a “results” page,
so just redirect to the permalink of the revision where we found that no
change was necessary; this is arguably confusing, but at least a bit
better than the previous behavior, which was to link to that revision as
a diff even if the edit was totally unrelated to Ranker.