Fix URI→statement ID conversion for forms and senses
Replacing the first hyphen with a dollar sign isn’t correct in those
cases; I think the easiest solution is to take advantage of the fact
that the GUID always has a constant length (32+4 characters), and split
the string based on that.

Previously, batch edits to form or sense statements would erroneously be
skipped as “nothing to do”, because we ended up looking for their
statement IDs in the parent lexeme.