Abort if statement ID contains no dollar sign
If users copy statement IDs out of the query service into list mode
(instead of using query mode), they might forget to replace the right
hyphen with a dollar sign when turning the statement URIs into statement
IDs. Raise an error with an explanatory message in that case, rather
than later crashing with an internal server error when we try to ask
wbgetentities for the whole statement ID as an entity ID and get an
error response. (I’ve tested that flask.abort() HTML-escapes the
message, of course.)