Force formatted entities and values to be in English
See T345881#9151012 – without this, the result depends on the user
language, so if we don’t specify a particular language, we need to split
the cache by the language somehow (or not cache results at all). But
using flask.g, as suggested in that comment, feels wrong in wbformat
(which otherwise doesn’t use Flask), and I don’t want to make a separate
API request to get the language associated with the session. So let’s
just use English, like for the rest of the tool, I think. (Maybe it’ll
become translatable eventually, no idea.)

Bug: T345881