Format Wikidata values/entities in user language
Note that it’s possible for the wbformat output and the surrounding
message to not match in language. (This isn’t new with this commit; I
just didn’t notice it before.) We might want to wrap the wbformat output
in a span to make sure it’s always tagged in the user language, even if
the surrounding message fell back to another language. But I think I’ll
look at this later.