Make batch mode input labels translatable
For the list modes, this is straightforward. For the query modes, we
restructure how to show the query service; I want to avoid a “lego
message” like “$1 query, selecting…” (in German, for instance, I think
the $1 could conceivably be in the accusative case: „Query/Anfrage für
den Wikidata Query Service…“), so get rid of query_service_name()
entirely and instead bake the name fully into the message. If we ever
get support for a second query service again (recall that support for
WCQS was removed in 561450bdbd due to the required authentication; see
T297995), I’ll just make a second copy of the two messages and let
translators copy+paste the invariant parts themselves.