Format reasons for rank using localized labels
This shows them in the interface language, like other mentioned
entities. To simplify the template code, put the lists of item IDs into
app.py, which also means we can test that the entities aren’t deleted;
do the prefetching in reason.html though (via a newly exposed template
global) so that we don’t have to duplicate it in each batch mode
function, and also so it’s closer to the place where the entities are
rendered.

This looks up (and tests) the reasons separately for Wikidata and
Commons, including a split cache in wbformat; I’ll fix that in a
separate commit.