Make rest of reason.html translatable
And with that, I think all the text in the tool is translatable! \o/

(Barring the two hard-coded error messages in app.py that have a TODO to
improve them anyway: 'CSRF error' and 'not logged in'. And also barring
no-oauth-request-token.html which only I or other developers should ever
see anyways. Maybe I’ll get around to these eventually but they’re not
blockers for starting to collect actual translations.)