Improve error handling in a few places
If people pasted just a statement ID into the “individual list” mode, it
would crash because the tuple returned by re.split() only contained
three elements, not four; tolerate this (we’ll get an error back from
the API, “The rank is not a valid rank”, but we handle that error well
and the rest of the batch can still proceed).

Also, replace the assertions for the correct SPARQL result variables
with a simple flask.abort(). We can even make the returned message nice
while we’re at it.